magnet kinda implemented
after a huge amount of time and effort and a 3-day period of my laptop not even posting so I couldn't work on this fucking mechanic, the magnet is in. I believe the camera flip and the switching hitbox positions, up direction and such works perfectly, but sometimes you don't properly land on surfaces and running/airtime is a little fucked atm. I don't think these will be as hard to do as the basic magnet implementation has been! (copium)